requirements.txt
pathspec
python-dotenv
openai

# The following libraries are also used (but should be included with python by default):
# os
# pathlib
# subprocess
# ast
#json